image-processor/requirements.txt
pika==1.2.0
requests==2.28.1
Pillow==9.2.0
